modified + comments

+++ src/s21_ceil.c
@@ -3,7 +3,7 @@
 long double s21_ceil(double x) {
     long double result = (long long int)x;
     if (s21_isnan(x) || s21_isinf(x)) {
-        result = x;
+        result = (long double)x;
     } else if (x != result && x > 0) { 
         result += 1;
     }
